refactor: clean up settings file for better structure
